requirements.txt
google-cloud
google-cloud-aiplatform
google-cloud-automl
google-cloud-vision
google-cloud-storage
google-cloud-pubsub